Force change for documents types set to nil
Because we're using a Definition class that is not AF::Base, we get peculiar errors, such as when you set a Definition property
to nil, the change is not recognized in the AF stack. Here we are forcing the change notification so we can successfully
remove any document type that has been set.