Correcting inverted attachment relationships, #1708
The initial implementation of attachments was incorrect and actually reflected a reverse relationship.

This is an intermediate fix. By simply changing the predicate to be attachmentOf, the relationship as it is coded is now correct.
This also prefigures upcoming work in which all relationships will be reversed from one-containing-many to one-part-of-other.
In other words, everything is going to become something-somethingOf, so why not just do this here and now with attachments.

The major issue to be aware of here is that none of the method names have changed, so the original, incorrect relationship
is still preserved in the names of the methods. This will be corrected in the future commit that reverses all the other
relationships.