Disable auto uploading for jQuery Uploader, ref #1768
Uploading several files at once was causing errors with the upload resulting in blank lines in the table of uploaded files.

This disables the auto-uploading feature in the jQuery uploader plugin resulting in files being queued instead of automatically
uploaded. A button is provided to complete the upload process for all files to save the trouble of having to click on each one.