Enable batch edit facility, ref #920
Assets need to be able to be edited in batches.

Enables editing of certain fields for a group of assets. Additionally, provides the ability to apply the same permissions
to a group of assets, including user and group access, as well as visibility.

Includes some refactoring of existing feature tests, and support for loading test objects in Solr only. This can drastically
reduce test time because Fedora is not required to generate the complete resource.