Ensure that Solr and Fedora records align for authority lists
Extra solr records may exist for authorities that are no longer present in Fedora.

In order to ensure that both Solr and Fedora have the exact same number of authority resources, all List and ListItem
records are deleted from Solr prior to loading the lists. The ListManager only uses Fedora-direct query methods
with the exact id, so Solr is never used while the new lists are updated.