Display text assets using the universal viewer, #1991
Not all text assets were being displayed by the UV because it was only
determined by mime type and not asset type.

Text assets that aren't pdf will be converted to pdf, making display via
the UV possible. The change here is to key off of asset rdf type.

Some helper methods were added to AssetPresenter as well as a
refactoring of ManifestService to account for the new changes.