Process pdf still image assets as images
If a pdf was uploaded as a still image asset, it was getting processed
as a pdf even though it should be processed as an image, with jp2
derivates created instead additional pdfs.

FileSet#create_derivatives is refactored to use the asset's type as the
logic for determining processing.

This clarifies the code considerably and there were additional
refactorings that came from it.