Refactor InboundRelationshipManagementService to use actor, ref #2042
When adding representations via the relationships tab of the asset's
edit page, the preferred representation was not being automatically
added if it wasn't present.

The service that manages relationships between assets and non-assets
during the asset's update and create process needed to be refactored to
use the CitiResourceActor. This actor will add a preferred
representation to a non-asset, and notify CITI accordingly.