Enable creation of single assets, ref #1806
Assets need to be created singly, just as they are in batches.

The UI for creating single assets is exactly the same as in batch.
The only difference being an intermediate file set may be selected.
The workflow is expanded so that an intermediate must be chosen in order
for the form to be submitted.