Refactor and update abilities with visibility, ref #1858
Visibility levels were not hierarchical and department visibility did
not include the ability of department users to edit assets, only to read
them.

Visibility permissions are interpreted in both Ability and via users and
groups, as per hydra-access-controls. This avoids a situation where the
specific permissions of a visibility setting can change in the
application code, but have not been updated in the users and groups to
reflect the new permissions.

The impact of this change is that ACLs in Fedora will need to manually
changed to match the new edit permission for users in the same
department as the asset's creating department. The creating
department will have to be added to all other assets.

Additional refactoring was made to indexers so that depositor metadata
such as depositor and department were indexed uniformly on those objects
that required it.