Enable converted pdfs to display in UV
Pdfs that are uploaded as images need to have their dimension metadata
taken from the convert jp2. This allows the UV to properly display them.
Otherwise, pdfs by themselves have no dimension information.

When the file set is indexed, if it contains a pdf that's been converted
to a jp2, fits will extract the dimension metadata from the jp2 and add
that to the solr document.

If no jp2 is present, or the file set doesn't contain any converted
pdfs, the dimension metadata remains unchanged.