Prevent metadata from accreting on fcr:metadata, ref #1416
When new versions of binaries were uploaded to existing ones, via the
FileSet UI edit page, the technical metadata extracted from the new
versions was being added to the old, existing metadata instead of
replacing it.

The problem is fixed by overriding Hydra::Work::CharacterizationService
so that it replaces the existing metadata instead of appending to it.